chore: 拆分value-listener
