+++ packages/zent/src/form/formulr/index.ts
@@ -11,7 +11,7 @@ export * from './models';
 export * from './validate';
 export * from './context';
 export * from './hooks';
-export * from './value-listener';
+export * from './listeners';
 export * from './maybe';
 export {
   UnknownFieldSetModelChildren,
